Fixed some compiler warnings.
git-svn-id: file:///home/nick/svnrepo/repo/indigo/trunk@3097 20aa11e4-6e6f-594c-a032-6291ace9ee65

+++ opencl/OpenCL.cpp
@@ -385,7 +385,7 @@ void OpenCL::queryDevices()
 			gpuDeviceInfo di;
 			di.device_number = current_device_number;
 			di.device_name = device_name_;
-			di.memory_size = (uint64)device_global_mem_size;
+			di.memory_size = (size_t)device_global_mem_size;
 			di.core_count = device_max_compute_units;
 			di.core_clock = device_max_clock_frequency; // in MHz
 			di.subsystem = Indigo::GPUDeviceSettings::SUBSYSTEM_OPENCL;
